voting + comments

+++ src/main/java/net/cefacem/app/dao/PostDAO.java
@@ -2,6 +2,6 @@ package net.cefacem.app.dao;
 
 import net.cefacem.app.model.Post;
 
-public interface PostDAO extends GenericDAO<Post, Integer> {
+public interface PostDAO extends GenericDAO<Post, Long> {
 
 }
