+++ src/main/java/net/cefacem/app/dao/UserDAO.java
@@ -2,7 +2,7 @@ package net.cefacem.app.dao;
 
 import net.cefacem.app.model.User;
 
-public interface UserDAO extends GenericDAO<User, Integer> {
+public interface UserDAO extends GenericDAO<User, Long> {
 	
 	public User findByUsername(String username);
 }
